# contracts/openapi_stub.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WAYMARK API",
    "description": "Explainable road risk, cell by cell, plus the action loop: region reports, an assistant, tracked measures with photo evidence and a verified-only adjusted risk estimate. Decision support only; human experts decide.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/regions/resolve": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "Resolve",
        "description": "Pure read: the name, cell count and bounds of a region. Public, like the other map reads.",
        "operationId": "resolve_api_regions_resolve_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "title": "Kind"
            }
          },
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300,
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/effects": {
      "get": {
        "tags": [
          "measures"
        ],
        "summary": "Effects",
        "description": "Categories, placeholder weights and the credit cap (read-only view of backend/config/effects.json).",
        "operationId": "effects_api_effects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/regions/{kind}/{key}/progress": {
      "get": {
        "tags": [
          "measures"
        ],
        "summary": "Progress",
        "operationId": "progress_api_regions__kind___key__progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Progress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/measures": {
      "post": {
        "tags": [
          "measures"
        ],
        "summary": "Create Measure",
        "operationId": "create_measure_api_measures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Measure"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/measures/{measure_id}": {
      "get": {
        "tags": [
          "measures"
        ],
        "summary": "Read Measure",
        "operationId": "read_measure_api_measures__measure_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "measure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Measure Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "measures"
        ],
        "summary": "Patch Measure",
        "operationId": "patch_measure_api_measures__measure_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "measure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Measure Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "measures"
        ],
        "summary": "Cancel Measure",
        "description": "Soft delete: the measure is marked cancelled and drops out of both percentages and the adjusted index.",
        "operationId": "cancel_measure_api_measures__measure_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "measure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Measure Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/measures/{measure_id}/comments": {
      "post": {
        "tags": [
          "measures"
        ],
        "summary": "Comment",
        "operationId": "comment_api_measures__measure_id__comments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "measure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Measure Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/measures/{measure_id}/verify": {
      "post": {
        "tags": [
          "measures"
        ],
        "summary": "Verify",
        "operationId": "verify_api_measures__measure_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "measure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Measure Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentCreate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CommentCreate"
      },
...
      "EffectsOut": {
        "properties": {
          "placeholder_weights": {
            "type": "boolean",
            "title": "Placeholder Weights"
          },
          "credit_cap": {
            "type": "number",
            "title": "Credit Cap"
          },
          "stage_factors": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Stage Factors"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryOut"
            },
            "type": "array",
            "title": "Categories"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "These are estimates from historical crash records, not predictions. The adjusted index is an overlay estimate, not a re-run of the risk model."
          }
        },
        "type": "object",
        "required": [
          "placeholder_weights",
          "credit_cap",
          "stage_factors",
          "categories"
        ],
        "title": "EffectsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Measure": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "region_kind": {
            "type": "string",
            "title": "Region Kind"
          },
          "region_key": {
            "type": "string",
            "title": "Region Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "category_label": {
            "type": "string",
            "title": "Category Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/MeasureStatus"
          },
          "effect_weight": {
            "type": "number",
            "title": "Effect Weight"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "cell_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cell Ids"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count",
            "default": 0
          },
          "evidence_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Required",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_id",
          "region_kind",
          "region_key",
          "title",
          "category",
          "category_label",
          "description",
          "owner_role",
          "status",
          "effect_weight",
          "source",
          "created_by",
          "created_at",
          "updated_at",
          "cell_ids"
        ],
        "title": "Measure"
      },
      "MeasureCreate": {
        "properties": {
          "region_kind": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Region Kind"
          },
          "region_key": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Region Key"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "category": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Category"
          },
          "cell_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Cell Ids"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "chat",
              "manual"
            ],
            "title": "Source",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "region_kind",
          "region_key",
          "title",
          "category",
          "cell_ids"
        ],
        "title": "MeasureCreate"
      },
      "MeasureDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "region_kind": {
            "type": "string",
            "title": "Region Kind"
          },
          "region_key": {
            "type": "string",
            "title": "Region Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "category_label": {
            "type": "string",
            "title": "Category Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/MeasureStatus"
          },
          "effect_weight": {
            "type": "number",
            "title": "Effect Weight"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "cell_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cell Ids"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count",
            "default": 0
          },
          "evidence_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Required",
            "default": []
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/MeasureEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "allowed_next": {
            "items": {
              "$ref": "#/components/schemas/MeasureStatus"
            },
            "type": "array",
            "title": "Allowed Next"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_id",
          "region_kind",
          "region_key",
          "title",
          "category",
          "category_label",
          "description",
          "owner_role",
          "status",
          "effect_weight",
          "source",
          "created_by",
          "created_at",
          "updated_at",
          "cell_ids",
          "events",
          "evidence",
          "allowed_next"
        ],
        "title": "MeasureDetail"
      },
...
      "MeasurePatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeasureStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "MeasurePatch"
      },
...
      "Progress": {
        "properties": {
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "measures": {
            "items": {
              "$ref": "#/components/schemas/Measure"
            },
            "type": "array",
            "title": "Measures"
          },
          "implementation_pct": {
            "type": "number",
            "title": "Implementation Pct"
          },
          "verified_pct": {
            "type": "number",
            "title": "Verified Pct"
          },
          "base_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Index"
          },
          "adjusted_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjusted Index"
          },
          "credit_cap": {
            "type": "number",
            "title": "Credit Cap"
          },
          "placeholder_weights": {
            "type": "boolean",
            "title": "Placeholder Weights"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/RiskSnapshotOut"
            },
            "type": "array",
            "title": "Snapshots"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "These are estimates from historical crash records, not predictions. The adjusted index is an overlay estimate, not a re-run of the risk model."
          }
        },
        "type": "object",
        "required": [
          "region",
          "measures",
          "implementation_pct",
          "verified_pct",
          "base_index",
          "adjusted_index",
          "credit_cap",
          "placeholder_weights",
          "snapshots"
        ],
        "title": "Progress"
      },
...
      "Region": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cell_count": {
            "type": "integer",
            "title": "Cell Count"
          },
          "bounds": {
            "$ref": "#/components/schemas/Bounds"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "key",
          "name",
          "cell_count",
          "bounds"
        ],
        "title": "Region"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "planner",
          "engineer",
          "community"
        ],
        "title": "Role"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}